MainMenu + Connection Changes
I made some changes to the main menu to include a list of rooms to make it easier for people to find the rooms they want, to do this I had to alter the entire way the project connects to rooms and had to change scenes using photons way of changing scenes instead of Unity's. Other changes if I remember correctly since last time include:

Teleporting
Drawable whiteboard
Changes to Textures on the walls and objects
Fixed the Avatars for the new ready player me SDK, this needs to be completed however

Next up is more features and now that the connection is rearranged it will be possible to send more information across to rooms easier so choosing avatars could work better and Usernames.